Resolve package tangles in Kotlin server package
