+++ config/src/main/kotlin/org/springframework/security/config/web/server/ServerContentSecurityPolicyDsl.kt
@@ -14,10 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.config.web.server.headers
-
-import org.springframework.security.config.web.server.ServerHttpSecurity
-import org.springframework.security.config.web.server.ServerSecurityMarker
+package org.springframework.security.config.web.server
 
 /**
  * A Kotlin DSL to configure the [ServerHttpSecurity] Content-Security-Policy header using
