+++ config/src/main/kotlin/org/springframework/security/config/web/server/ServerXssProtectionDsl.kt
@@ -14,40 +14,30 @@
  * limitations under the License.
  */
 
-package org.springframework.security.config.web.server.headers
-
-import org.springframework.security.config.web.server.ServerHttpSecurity
-import org.springframework.security.config.web.server.ServerSecurityMarker
-import org.springframework.security.web.server.header.XFrameOptionsServerHttpHeadersWriter
+package org.springframework.security.config.web.server
 
 /**
- * A Kotlin DSL to configure the [ServerHttpSecurity] X-Frame-Options header using
+ * A Kotlin DSL to configure the [ServerHttpSecurity] XSS protection header using
  * idiomatic Kotlin code.
  *
  * @author Eleftheria Stein
  * @since 5.4
- * @property mode the X-Frame-Options mode to set in the response header.
  */
 @ServerSecurityMarker
-class ServerFrameOptionsDsl {
-    var mode: XFrameOptionsServerHttpHeadersWriter.Mode? = null
-
+class ServerXssProtectionDsl {
     private var disabled = false
 
     /**
-     * Disables the X-Frame-Options response header
+     * Disables cache control response headers
      */
     fun disable() {
         disabled = true
     }
 
-    internal fun get(): (ServerHttpSecurity.HeaderSpec.FrameOptionsSpec) -> Unit {
-        return { frameOptions ->
-            mode?.also {
-                frameOptions.mode(mode)
-            }
+    internal fun get(): (ServerHttpSecurity.HeaderSpec.XssProtectionSpec) -> Unit {
+        return { xss ->
             if (disabled) {
-                frameOptions.disable()
+                xss.disable()
             }
         }
     }
